Fixed switch statement label error
Declarations cannot immediately follow a label (like the case part of a switch statement). This can be fixed by either adding a semicolon after the case or surrounding the code block associated with a case in curly braces, the latter was chosen as it was cleaner. 

Also added parenthesis to incorrect method calls